+++ bin/miden-cli/src/info.rs
@@ -1,19 +1,10 @@
 use std::fs;
 
-use miden_client::{
-    auth::TransactionAuthenticator,
-    crypto::FeltRng,
-    rpc::NodeRpcClient,
-    store::{NoteFilter, Store},
-    Client,
-};
+use miden_client::{crypto::FeltRng, store::NoteFilter, Client};
 
 use super::config::CliConfig;
 
-pub fn print_client_info<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator>(
-    client: &Client<N, R, S, A>,
-    config: &CliConfig,
-) -> Result<(), String> {
+pub fn print_client_info(client: &Client<impl FeltRng>, config: &CliConfig) -> Result<(), String> {
     println!("Client version: {}", env!("CARGO_PKG_VERSION"));
     print_config_stats(config)?;
     print_client_stats(client)
@@ -21,9 +12,7 @@ pub fn print_client_info<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionA
 
 // HELPERS
 // ================================================================================================
-fn print_client_stats<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator>(
-    client: &Client<N, R, S, A>,
-) -> Result<(), String> {
+fn print_client_stats(client: &Client<impl FeltRng>) -> Result<(), String> {
     println!("Block number: {}", client.get_sync_height().map_err(|e| e.to_string())?);
     println!(
         "Tracked accounts: {}",