+++ crates/rust-client/src/notes/import.rs
@@ -4,16 +4,14 @@ use miden_objects::{
     crypto::rand::FeltRng,
     notes::{Note, NoteDetails, NoteFile, NoteId, NoteInclusionProof, NoteMetadata, NoteTag},
 };
-use miden_tx::auth::TransactionAuthenticator;
 use winter_maybe_async::maybe_await;
 
 use crate::{
-    rpc::NodeRpcClient,
-    store::{ExpectedNoteState, InputNoteRecord, Store},
+    store::{ExpectedNoteState, InputNoteRecord},
     Client, ClientError,
 };
 
-impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client<N, R, S, A> {
+impl<R: FeltRng> Client<R> {
     // INPUT NOTE CREATION
     // --------------------------------------------------------------------------------------------