+++ crates/rust-client/src/store/web_store/mod.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "async")]
+use alloc::boxed::Box;
 use alloc::{collections::BTreeMap, vec::Vec};
 
 use miden_objects::{
@@ -8,13 +10,13 @@ use miden_objects::{
 };
 use wasm_bindgen::prelude::*;
 use wasm_bindgen_futures::*;
-use winter_maybe_async::{maybe_async, maybe_await};
+use winter_maybe_async::*;
 
+use super::{
+    ChainMmrNodeFilter, InputNoteRecord, NoteFilter, OutputNoteRecord, Store, StoreError,
+    TransactionFilter,
+};
 use crate::{
-    store::{
-        ChainMmrNodeFilter, InputNoteRecord, NoteFilter, OutputNoteRecord, Store, StoreError,
-        TransactionFilter,
-    },
     sync::StateSyncUpdate,
     transactions::{TransactionRecord, TransactionResult},
 };
@@ -40,34 +42,33 @@ impl WebStore {
         Ok(WebStore {})
     }
 }
-
+#[maybe_async_trait]
 impl Store for WebStore {
     // SYNC
     // --------------------------------------------------------------------------------------------
-
     #[maybe_async]
     fn get_note_tags(&self) -> Result<Vec<NoteTag>, StoreError> {
-        maybe_await!(self.get_note_tags())
+        self.get_note_tags().await
     }
 
     #[maybe_async]
     fn add_note_tag(&self, tag: NoteTag) -> Result<bool, StoreError> {
-        maybe_await!(self.add_note_tag(tag))
+        self.add_note_tag(tag).await
     }
 
     #[maybe_async]
     fn remove_note_tag(&self, tag: NoteTag) -> Result<bool, StoreError> {
-        maybe_await!(self.remove_note_tag(tag))
+        self.remove_note_tag(tag).await
     }
 
     #[maybe_async]
     fn get_sync_height(&self) -> Result<u32, StoreError> {
-        maybe_await!(self.get_sync_height())
+        self.get_sync_height().await
     }
 
     #[maybe_async]
     fn apply_state_sync(&self, state_sync_update: StateSyncUpdate) -> Result<(), StoreError> {
-        maybe_await!(self.apply_state_sync(state_sync_update))
+        self.apply_state_sync(state_sync_update).await
     }
 
     // TRANSACTIONS
@@ -78,28 +79,27 @@ impl Store for WebStore {
         &self,
         transaction_filter: TransactionFilter,
     ) -> Result<Vec<TransactionRecord>, StoreError> {
-        maybe_await!(self.get_transactions(transaction_filter))
+        self.get_transactions(transaction_filter).await
     }
 
     #[maybe_async]
     fn apply_transaction(&self, tx_result: TransactionResult) -> Result<(), StoreError> {
-        maybe_await!(self.apply_transaction(tx_result))
+        self.apply_transaction(tx_result).await
     }
 
     // NOTES
     // --------------------------------------------------------------------------------------------
-
     #[maybe_async]
-    fn get_input_notes(&self, filter: NoteFilter<'_>) -> Result<Vec<InputNoteRecord>, StoreError> {
-        maybe_await!(self.get_input_notes(filter))
+    fn get_input_notes(&self, filter: NoteFilter) -> Result<Vec<InputNoteRecord>, StoreError> {
+        self.get_input_notes(filter).await
     }
 
     #[maybe_async]
     fn get_output_notes(
         &self,
-        note_filter: NoteFilter<'_>,
+        note_filter: NoteFilter,
     ) -> Result<Vec<OutputNoteRecord>, StoreError> {
-        maybe_await!(self.get_output_notes(note_filter))
+        self.get_output_notes(note_filter).await
     }
 
     #[maybe_async]
@@ -117,7 +117,7 @@ impl Store for WebStore {
         chain_mmr_peaks: MmrPeaks,
         has_client_notes: bool,
     ) -> Result<(), StoreError> {
-        maybe_await!(self.insert_block_header(block_header, chain_mmr_peaks, has_client_notes))
+        self.insert_block_header(block_header, chain_mmr_peaks, has_client_notes).await
     }
 
     #[maybe_async]
@@ -125,30 +125,30 @@ impl Store for WebStore {
         &self,
         block_numbers: &[u32],
     ) -> Result<Vec<(BlockHeader, bool)>, StoreError> {
-        maybe_await!(self.get_block_headers(block_numbers))
+        self.get_block_headers(block_numbers).await
     }
 
     #[maybe_async]
     fn get_tracked_block_headers(&self) -> Result<Vec<BlockHeader>, StoreError> {
-        maybe_await!(self.get_tracked_block_headers())
+        self.get_tracked_block_headers().await
     }
 
     #[maybe_async]
-    fn get_chain_mmr_nodes<'a>(
+    fn get_chain_mmr_nodes(
         &self,
-        filter: ChainMmrNodeFilter<'a>,
+        filter: ChainMmrNodeFilter,
     ) -> Result<BTreeMap<InOrderIndex, Digest>, StoreError> {
-        maybe_await!(self.get_chain_mmr_nodes(filter))
+        self.get_chain_mmr_nodes(filter).await
     }
 
     #[maybe_async]
     fn insert_chain_mmr_nodes(&self, nodes: &[(InOrderIndex, Digest)]) -> Result<(), StoreError> {
-        maybe_await!(self.insert_chain_mmr_nodes(nodes))
+        self.insert_chain_mmr_nodes(nodes).await
     }
 
     #[maybe_async]
     fn get_chain_mmr_peaks_by_block_num(&self, block_num: u32) -> Result<MmrPeaks, StoreError> {
-        maybe_await!(self.get_chain_mmr_peaks_by_block_num(block_num))
+        self.get_chain_mmr_peaks_by_block_num(block_num).await
     }
 
     // ACCOUNTS
@@ -161,43 +161,12 @@ impl Store for WebStore {
         account_seed: Option<Word>,
         auth_info: &AuthSecretKey,
     ) -> Result<(), StoreError> {
-        maybe_await!(self.insert_account(account, account_seed, auth_info))
+        self.insert_account(account, account_seed, auth_info).await
     }
 
     #[maybe_async]
     fn get_account_ids(&self) -> Result<Vec<AccountId>, StoreError> {
-        maybe_await!(self.get_account_ids())
-    }
-
-    #[maybe_async]
-    fn get_account_headers(&self) -> Result<Vec<(AccountHeader, Option<Word>)>, StoreError> {
-        maybe_await!(self.get_account_headers())
-    }
-
-    #[maybe_async]
-    fn get_account_header(
-        &self,
-        account_id: AccountId,
-    ) -> Result<(AccountHeader, Option<Word>), StoreError> {
-        maybe_await!(self.get_account_header(account_id))
-    }
-
-    #[maybe_async]
-    fn get_account_header_by_hash(
-        &self,
-        account_hash: Digest,
-    ) -> Result<Option<AccountHeader>, StoreError> {
-        maybe_await!(self.get_account_header_by_hash(account_hash))
-    }
-
-    #[maybe_async]
-    fn get_account(&self, account_id: AccountId) -> Result<(Account, Option<Word>), StoreError> {
-        maybe_await!(self.get_account(account_id))
-    }
-
-    #[maybe_async]
-    fn get_account_auth(&self, account_id: AccountId) -> Result<AuthSecretKey, StoreError> {
-        maybe_await!(self.get_account_auth(account_id))
+        self.get_account_ids().await
     }
 
     fn get_account_auth_by_pub_key(&self, pub_key: Word) -> Result<AuthSecretKey, StoreError> {
@@ -205,7 +174,38 @@ impl Store for WebStore {
     }
 
     #[maybe_async]
+    fn get_account_headers(&self) -> Result<Vec<(AccountHeader, Option<Word>)>, StoreError> {
+        self.get_account_headers().await
+    }
+
+    #[maybe_async]
+    fn get_account_header(
+        &self,
+        account_id: AccountId,
+    ) -> Result<(AccountHeader, Option<Word>), StoreError> {
+        self.get_account_header(account_id).await
+    }
+
+    #[maybe_async]
+    fn get_account_header_by_hash(
+        &self,
+        account_hash: Digest,
+    ) -> Result<Option<AccountHeader>, StoreError> {
+        self.get_account_header_by_hash(account_hash).await
+    }
+
+    #[maybe_async]
+    fn get_account(&self, account_id: AccountId) -> Result<(Account, Option<Word>), StoreError> {
+        self.get_account(account_id).await
+    }
+
+    #[maybe_async]
+    fn get_account_auth(&self, account_id: AccountId) -> Result<AuthSecretKey, StoreError> {
+        self.get_account_auth(account_id).await
+    }
+
+    #[maybe_async]
     fn get_unspent_input_note_nullifiers(&self) -> Result<Vec<Nullifier>, StoreError> {
-        maybe_await!(self.get_unspent_input_note_nullifiers())
+        self.get_unspent_input_note_nullifiers().await
     }
 }